requirements.txt
# ======================================================
# CORE - Google ADK
# ======================================================
google-adk==1.14.0

# ======================================================
# Google Cloud / Auth (needed for MCP + Vertex AI)
# ======================================================
google-auth
google-auth-httplib2
google-cloud-aiplatform
google-cloud-logging

# ======================================================
# Tavily Search (IMPORTANT - you are using this)
# ======================================================
langchain-community
tavily-python

# ======================================================
# Env support
# ======================================================
python-dotenv

# ======================================================
# HTTP requests
# ======================================================
requests

# ======================================================
# Optional (only if you use scraping later)
# ======================================================
beautifulsoup4
lxml
